Remove incorrect validations
The length of the format string does not necessarily match the length of
the actual date, so this validation is incorrect.

I'm not sure when this validation is useful anyways (if some malformed
date comes in the field from the server I guess).

In my opinion, the easiest way to fix this is to remove the validation.
Users wanting client side validations in foundation should use
`foundation-abide`.